refactor: move isOrg function (#15012)
