+++ frontend/dashboard/utils/orgUtils/index.ts
@@ -1,0 +1,1 @@
+export { isOrg } from './orgUtils';
